refactor(core): extract transaction types to transaction.rs
Move Transaction and TxOut types to dedicated transaction module.
